LUN-265: Add bottom tabs navigator
