+++ assets/images/index.ts
@@ -5,6 +5,8 @@ import ArrowRightIcon from './arrow-right-icon.svg'
 import BannerGreen from './banner-green.png'
 import BannerRed from './banner-red.png'
 import BannerYellow from './banner-yellow.png'
+import BookOutlineIconGrey from './book-outline-icon-grey.svg'
+import BookOutlineIconWhite from './book-outline-icon-white.svg'
 import CheckCircleIconBold from './check-circle-icon-bold.svg'
 import CheckCircleIconGreen from './check-circle-icon-green.svg'
 import CheckCircleIconWhite from './check-circle-icon-white.svg'
@@ -25,8 +27,12 @@ import HeaderCircleIcon from './header-cirlce-icon.svg'
 import HeaderLinesIcon from './header-lines-icon.svg'
 import HeaderSquareIcon from './header-square-icon.svg'
 import HeaderStarIcon from './header-star-icon.svg'
+import HeartOutlineIconGrey from './heart-outline-icon-grey.svg'
+import HeartOutlineIconWhite from './heart-outline-icon-white.svg'
 import HomeCircleIconBlue from './home-circle-icon-blue.svg'
 import HomeCircleIconWhite from './home-circle-icon-white.svg'
+import HomeOutlineIconGrey from './home-outline-icon-grey.svg'
+import HomeOutlineIconWhite from './home-outline-icon-white.svg'
 import InfoCircleIcon from './info-circle-icon.svg'
 import ListIcon from './list-icon.svg'
 import LunesIcon from './lunes-icon.svg'
@@ -35,6 +41,8 @@ import OpenLockIcon from './open-lock-icon.svg'
 import QRCodeIcon from './qr-code-icon.svg'
 import RepeatIcon from './repeat-icon.svg'
 import ShareIcon from './share-icon.svg'
+import StarOutlineIconGrey from './star-outline-icon-grey.svg'
+import StarOutlineIconWhite from './star-outline-icon-white.svg'
 import TrashIcon from './trash-bin-icon.svg'
 import TrophyIcon from './trophy-icon.svg'
 import VolumeUpCircleIcon from './volume-up-circle-icon.svg'
@@ -47,6 +55,8 @@ export {
   BannerGreen,
   BannerRed,
   BannerYellow,
+  BookOutlineIconGrey,
+  BookOutlineIconWhite,
   CheckCircleIcon,
   CheckCircleIconBold,
   CheckCircleIconWhite,
@@ -67,17 +77,23 @@ export {
   HeaderLinesIcon,
   HeaderSquareIcon,
   HeaderStarIcon,
+  HeartOutlineIconGrey,
+  HeartOutlineIconWhite,
   HomeCircleIconBlue,
   HomeCircleIconWhite,
+  HomeOutlineIconGrey,
+  HomeOutlineIconWhite,
   InfoCircleIcon,
   ListIcon,
   LunesIcon,
   MenuIcon,
+  OpenLockIcon,
   QRCodeIcon,
   RepeatIcon,
   ShareIcon,
+  StarOutlineIconGrey,
+  StarOutlineIconWhite,
   TrashIcon,
   TrophyIcon,
-  VolumeUpCircleIcon,
-  OpenLockIcon
+  VolumeUpCircleIcon
 }
